refactor : 상단 Controller에서 예외를 잡도록 수정
